Move wcif helper to wcif util
