--- tnoodle-ui/src/main/helper/wcif.helper.js
@@ -1,15 +1,0 @@
-export const copiesExtensionId =
-    "org.worldcubeassociation.tnoodle.SheetCopyCount";
-/**
- * This is the default extension object the backend expects
- * @param {} copies
- */
-export const getDefaultCopiesExtension = () => {
-    return {
-        id: copiesExtensionId,
-        specUrl: "",
-        data: {
-            numCopies: 1,
-        },
-    };
-};
